+++ src/org/openscience/cdk/test/reaction/type/RearrangementCation2ReactionTest.java
@@ -93,7 +93,7 @@ public class RearrangementCation2ReactionTest extends CDKTestCase {
         
         
         IMolecule product = setOfReactions.getReaction(0).getProducts().getMolecule(0);
-        Assert.assertEquals(1, product.getAtom(2).getFormalCharge());
+        Assert.assertEquals(1, product.getAtom(2).getFormalCharge().intValue());
         Assert.assertEquals(0, product.getConnectedLonePairsCount(molecule.getAtom(1)));
         
         /*C=C-[C+]-C*/
